+++ app/queries/slot_query.rb
@@ -13,7 +13,7 @@ module SlotQuery
     end
 
     # I don't like the split in direction and cursor because they belong together
-    def retrieve(filter: nil, moment: Time.zone.now, cursor: nil)
+    def retrieve(mode: nil, moment: Time.zone.now, cursor: nil)
       # query with cursor
       case @direction
       when 'before'
@@ -22,26 +22,24 @@ module SlotQuery
         @relation.where(after_cursor(cursor)).ordered
       else
         # query without cursor
-        case filter
+        case mode
         when nil
         when 'all'
           @relation
         when 'finished'
-          @relation.where(finished moment).ordered_rev
+          @relation.where(finished(moment)).ordered_rev
         when 'upcoming', 'ongoing', 'past', 'now'
-          # here we send the 'filter' as a message to this SlotQuery:OwnSlots
+          # here we send the 'mode' as a message to this SlotQuery:OwnSlots
           # class, which means, we are calling the method with the name of
-          # the 'filter'
-          @relation.where(send filter, moment).ordered
+          # the 'mode'
+          @relation.where(send(mode, moment)).ordered
         else
-          # TODO: make a helper for enriched airbrake error messages
-          error_string = "unknown pagination filter #{filter}"
+          error_string = "unknown pagination mode #{mode}"
           msg = { message: error_string }
           Rails.logger.error { error_string } unless Rails.env.test?
           Airbrake.notify(PaginationError, msg)
           fail PaginationError, msg if Rails.env.test? || Rails.env.development?
           # TODO: check if we should call 'new' for custom error classes? Why?
-          # fail PaginationError.new(msg) if Rails.env.test? || Rails.env.development?
         end
       end
     end