+++ spec/acceptance/users_spec.rb
@@ -422,11 +422,11 @@ resource "Users" do
                          "send, **status** and **moment** are ignored."
       parameter :after, "Pagination cursor to retrieve slots which do happen" \
                         " AFTER the slot represented by this cursor. If a " \
-                        "cursor is send, **filter** and **moment** are ignored."
+                        "cursor is send, **mode** and **moment** are ignored."
 
       response_field :paging, "Hash containing relevant paging parameters."
       response_field :limit, "Maximum number of slots returned."
-      response_field :filter, "Types of slots which were requested."
+      response_field :mode, "Types of slots which were requested."
       response_field :moment, "Point-in-time which was used for the query."
       response_field :before, "Cursor that represents the first item in the " \
                               "response dataset."
@@ -441,7 +441,7 @@ resource "Users" do
           friend
         end
         let(:id) { friend.id }
-        let(:filter) { 'upcoming' }
+        let(:mode) { 'upcoming' }
         let(:moment) { Time.zone.now.as_json }
         let(:limit) { 3 }
 
@@ -464,7 +464,7 @@ resource "Users" do
                       "If there are more than **limit** results, '*paging*' " \
                       "has **before** and **after** cursors which can be used" \
                       " for subsequent requests. The first request should " \
-                      "always be made with **filter** '*upcoming*' to make " \
+                      "always be made with **mode** '*upcoming*' to make " \
                       "sure no results are skipped." \
                       "'*data*' contains an array which includes " \
                       "StandardSlots & ReSlots\n\n" \
@@ -515,10 +515,10 @@ resource "Users" do
           json = JSON.parse(response_body)
 
           expect(json).to have_key 'paging'
-          expect(json['paging']).to have_key('filter')
+          expect(json['paging']).to have_key('mode')
           expect(json['paging']).to have_key('after')
           expect(json['paging']).to have_key('limit')
-          expect(json['paging']['filter']).to be nil
+          expect(json['paging']['mode']).to be nil
           expect(json['paging']['after']).to be nil
           expect(json['paging']['limit']).to eq PAGINATION_MAX_LIMIT
           # expect(response_body).to include(re_slots.first.title)